Fix incorrect zoom to
When the same point was the extreme in two directions, e.g. when looking onto a tilted rhombus where the topmost point was also the leftmost and the bottom-most was also the rightmost, the camera was incorrectly placed. This due to the same Point3d object being transformed twice when converting between camera and model space. Be careful with bang methods!